Little fixes.

+++ gui/helpers.cpp
@@ -35,7 +35,7 @@ void nt::DrawQueue::push(char z, sf::Drawable& object)
   queue_.insert(std::make_pair(z, &object));
 }
 
-short int nt::stringToInt(const std::string& string)
+int nt::stringToInt(const std::string& string)
 {
   return std::stoi(string);
 }
